Initial working Sidecar integration
The design is to check the Sidecar status and if
anything goes wrong in doing that, we leave the task up.
If all goes correctly and we can find out that the
container is either Unhealthy or Tombstoned, we tear it
down.

We back off for a fixed interval after starting
the container, we don't want to tear it down while it's
still deploying.