Attempt harder to only log errors as errors
Anything where we scrape the logs is going to be a compromise
and this attempts to push it more in the direction of only
logging actual errors as errors. Of course, some applications
will send errors to stderr and everything else to stdout as
a proper application should do, but these days it seems to
be increasingly rare that applications work like that. :(